Reorganized architecture
