+++ .gitignore
@@ -128,3 +128,7 @@ dist
 .yarn/build-state.yml
 .yarn/install-state.gz
 .pnp.*
+.idea/ElectivaIIJuanTobonHaroldPadillaHugoCastrillon.iml
+.idea/misc.xml
+.idea/modules.xml
+.idea/vcs.xml
