+++ src/core/domain/mocks/FollowingMocks.js
