+++ src/infrastructure/express/routes/AuthRoutes.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const { registerUser, loginUser } = require("../controllers/authControllers");
+const { registerUser, loginUser } = require("../../../adapters/controllers/authControllers");
 const { validateRegister, validateLogin } = require("../middlewares/middlewareAuth");
 
 router.post('/auth/register', validateRegister, registerUser);
